swap: Introduce a new umount-post.target for swap teardown
Previously, the swap disabling operation was linked to the umount
operation. This caused issues within system shutdown or reboot due to the
swapoff failing with not enough RAM present.

To reproduce this condition, a system with the following characteristics
can be used:

  RAM 2Gib
 Swap 4Gib

Steps required:

 # mount -t tmpfs -o size=4G nodev /mnt/tmpfs
 # dd if=/dev/zero of=/mnt/tmpfs/testfile bs=1G
 # systemctl reboot

Expected results:

 System restarts successfully

Actual results:

 Reboot operation times out due to the swapoff failing

To address this condition, this patch introduces a new umount-post.target
which is marked as a default dependency and conflict for swap units. It is
started activated after the umount.target has completed successfully and
local mount points, including tmpfs filesystems, have been unmounted.